+++ extensions/iceworks-project-creator/CHANGELOG.md
@@ -3,6 +3,7 @@
 # 0.3.18
 
 - refactor: material source card
+- feat: add custom scaffold 
 
 # 0.3.17